+++ src/data/users.ts
@@ -2,8 +2,8 @@ import { IUser } from "../types";
 
 export const users: IUser[] = [
   {
-    id: '123',
-    username: 'user',
+    id: 'a74c46a2-f3c7-425c-9788-314456b25c0e',
+    username: 'initial_user',
     age: 20,
     hobbies: []
   }
